Make admin panel responsive

+++ admin/src/components/Rooms.jsx
@@ -81,6 +81,9 @@ const Rooms = () => {
     }
 
     const deleteRoom = (room) => {
+        const userConfirmation = confirm(`Are you sure, you want to delete ${room?.roomName || "Room"}`);
+
+        if(!userConfirmation) return ;
         fetch(`${BASE_URL}/deleteRoom`, {
             "method": "delete",
             "body": JSON.stringify(room),
